fix(features): dead-channel std must be computed in float64
Real-data crash (2026-07-01 TU1, RAWTurbineVib__3): channels exactly
constant at -7.0 in every sample of every file, but VibFeaturizer's
batch std computed in the input's own float32 picks up pairwise-
summation rounding noise (~|c|*eps/2 = 4.77e-07 for c=-7.0) whose
presence depends on the batch's total element count -- above the 1e-9
dead threshold for some files of the stream, exactly 0.0 for others.
The same physically-dead channel flip-flopped live/dead across files,
changed the feature-row width mid-stream, and crashed
_extract_stream_features ("could not broadcast input array from shape
(36,) into shape (72,)"). In float64 a constant channel's std is
exactly 0.0 for every batch shape; genuinely live channels are
unaffected.

+++ src/rowii/signals/features.py
@@ -317,7 +317,19 @@ class VibFeaturizer:
 
         live_channels: list[int] = []
         for ch in range(n_channels):
-            std = float(windows[:, :, ch].std())
+            # float64, NOT the input's own float32: float32 pairwise summation
+            # injects rounding noise of ~|c| * eps/2 into the std of a channel
+            # exactly constant at c, and the noise's magnitude depends on the
+            # batch's total element count -- measured on real data (2026-07-01
+            # TU1 RAWTurbineVib__3, channels constant at -7.0): 4.77e-07 for some
+            # per-file batch shapes, exactly 0.0 for others, i.e. ABOVE the 1e-9
+            # dead threshold for some files of a stream and below it for the
+            # rest, flip-flopping the same physically-dead channel live/dead
+            # across files and changing the feature-row width mid-stream (which
+            # crashes _extract_stream_features' preallocated-matrix assignment).
+            # In float64 a constant channel's std is exactly 0.0 for every batch
+            # shape at any realistic batch size.
+            std = float(windows[:, :, ch].astype(np.float64).std())
             if std < _DEAD_CHANNEL_STD:
                 logger.warning(
                     "VibFeaturizer: channel %d is dead (std=%.3e < %.1e over the full "
